moved showQuestion function to utils/common
